requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose==3.3.0
python-multipart==0.0.6
python-dotenv==1.0.0
jinja2==3.1.2
PyJWT
python-jose[cryptography]
pydantic-settings
python-dotenv
jose